Fix wasm_bindgen_test tests
