+++ .github/workflows/test.yml
@@ -100,4 +100,4 @@ jobs:
     - name: Compile Rust code to WASM
       run: wasm-pack build crates/pluot --target web
     - name: Run unit tests
-      run: wasm-pack test crates/pluot --headless --chrome
+      run: wasm-pack test --headless --chrome crates/pluot
