+++ crates/pluot/src/lib.rs
@@ -5,6 +5,7 @@ mod plugins;
 // Export things needed for layer-based plotting via Rust.
 pub use pluot_core::params::{GraphicsFormat, ViewMode};
 pub use pluot_core::render_traits::{AspectRatioMode, UnitsMode, ViewParams, MarginParams};
+pub use pluot_core::{LayerParams as RawLayerParams, RenderParams as RawRenderParams};
 
 
 // Re-export layer param types for convenience.
@@ -20,15 +21,18 @@ pub use pluot_zarr::layers::zarr_point_3d_layer::ZarrPoint3dLayerParams;
 pub use pluot_zarr::layers::ome_zarr_bitmap_layer::OmeZarrBitmapLayerParams;
 pub use pluot_zarr::layers::ome_zarr_multiscale_layer::OmeZarrMultiscaleLayerParams;
 
+mod render_params;
+
+pub use crate::render_params::{RenderParams, LayerParams};
+
 // Unified exports.
+mod render;
+pub use crate::render::{render};
+
+// Exports for WASM bindings.
 #[cfg(target_arch = "wasm32")]
 pub use pluot_core::bindings::wasm::{render_wasm, set_panic_hook};
 
+// Exports for Python bindings.
 #[cfg(all(not(target_arch = "wasm32"), feature = "python"))]
 pub use pluot_core::bindings::python::{render_py};
-
-#[cfg(all(not(target_arch = "wasm32"), not(feature = "python")))]
-mod render;
-
-#[cfg(all(not(target_arch = "wasm32"), not(feature = "python")))]
-pub use crate::render::{render, RenderParams, LayerParams};
