+++ crates/pluot_core/src/lib.rs
@@ -30,6 +30,7 @@ pub use wgpu;
 pub use crate::params::{RenderParams, PlotParams, LayerParams, GraphicsFormat, LayeredPlotRenderParams, ViewMode};
 pub use crate::render_traits::{AspectRatioMode, UnitsMode, ViewParams, MarginParams};
 pub use crate::registry::{LayerRegistration, get_layer_from_registry};
+pub use crate::render::{render};
 pub use crate::picking::{pick, PickingResult, LayerPickingResult};
 pub use crate::viewport::{project, unproject, get_bounds, camera_matrix_to_zoom_and_translation};
